Ok now i fixed it all
